Product -상품 검색 기능중 없는 상품일때 기능 추가 -주석 추가

+++ src/main/java/com/example/commerce_back_office/controller/ReviewController.java
@@ -30,7 +30,7 @@ public class ReviewController {
     @PostMapping
     public ResponseEntity<ReviewResponseDto> createReview(
             @AuthenticationPrincipal CustomUserDetails userPrincipal,
-            @PathVariable int productId,
+            @PathVariable long productId,
             @Valid @RequestBody ReviewRequestDto request) {
 
         ReviewResponseDto response = reviewService.save(userPrincipal.getUser(), productId, request);
